% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_scorers.R
\name{create_table_scorers}
\alias{create_table_scorers}
\title{Create Table Scorer}
\usage{
create_table_scorers(url_games)
}
\arguments{
\item{url_games}{single url to game as retrieved by get_links_game()}
}
\value{
scorer table
}
\description{
Create Table Scorer
}
\examples{
games_u11 <- get_links_game()[7,]

create_table_scorers(url_games = games_u11$links_url)

}
