% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_saisonmanager_url.R
\name{default_saisonmanager_url}
\alias{default_saisonmanager_url}
\title{Default URL for FVBB Saisonmanager}
\usage{
default_saisonmanager_url()
}
\value{
url to FVBB saionsmanager (default: "https://fvbb.saisonmanager.de")
}
\description{
Default URL for FVBB Saisonmanager
}
\examples{
default_saisonmanager_url()
}
