% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_scorers.R
\name{get_game_master}
\alias{get_game_master}
\title{Get Game Master Data}
\usage{
get_game_master(game_link_full)
}
\arguments{
\item{game_link_full}{link to game stats url}
}
\value{
table with game master data
}
\description{
Get Game Master Data
}
\examples{
games_u11 <- get_links_game()[7,]
completed_games_u11 <- urls_completed_games(games_u11$links_url)
if(length(completed_games_u11)>0) get_game_master(completed_games_u11[1]) 
}
