% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_links.R
\name{get_links}
\alias{get_links}
\title{Get Links}
\usage{
get_links(url_saisonmanager = default_saisonmanager_url(),
  select = "*")
}
\arguments{
\item{url_saisonmanager}{base url to FVBB saisonmanager (default: default_saisonmanager_url())}

\item{select}{which type of info ("spielplan", "table, "scorer") to select (default: *, i.e. 
everything)}
}
\value{
tibble with links to spielplaene, tables and scorers
}
\description{
Get Links
}
\examples{
\dontrun{
get_links(url_saisonmanager = "https://fvbb.saisonmanager.de")
}
}
