% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_links.R
\name{get_links_game}
\alias{get_links_game}
\title{Get Links To Games}
\usage{
get_links_game()
}
\value{
tibble with links to games
}
\description{
Get Links To Games
}
\examples{
\dontrun{
get_links_game()
}
}
