% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_links.R
\name{get_links_scorer}
\alias{get_links_scorer}
\title{Get Links To Scorers}
\usage{
get_links_scorer(url_saisonmanager = default_saisonmanager_url())
}
\arguments{
\item{url_saisonmanager}{base url to FVBB saisonmanager (default: default_saisonmanager_url())}
}
\value{
tibble with links to scorers
}
\description{
Get Links To Scorers
}
\examples{
\dontrun{
get_links_scorer()
}
}
