% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_links.R
\name{get_links_table}
\alias{get_links_table}
\title{Get Links To Tables}
\usage{
get_links_table(url_saisonmanager = default_saisonmanager_url())
}
\arguments{
\item{url_saisonmanager}{base url to FVBB saisonmanager (default: default_saisonmanager_url())}
}
\value{
tibble with links to tables
}
\description{
Get Links To Tables
}
\examples{
\dontrun{
get_links_table()
}
}
