% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scorer.R
\name{get_scorers}
\alias{get_scorers}
\title{Get Scorers}
\usage{
get_scorers(game_link_full)
}
\arguments{
\item{game_link_full}{link to game stats url}
}
\value{
table with scorers
}
\description{
Get Scorers
}
\examples{
path_list <- list(
saisonmanager = "https://fvbb.saisonmanager.de",
spielplan =  "<saisonmanager>/index.php?seite=spielplan",
spielplan_u11 = "<spielplan>&table=789")

paths <- kwb.utils::resolve(path_list)

content <- xml2::read_html(paths$spielplan_u11)

game_results_links <- xml2::read_html(paths$spielplan_u11) \%>\% 
rvest::html_nodes("td.center a")

is_over <- rvest::html_text(game_results_links) != ""

game_links <- game_results_links[is_over] \%>\%  rvest::html_attr("href") 
game_links_full <- sprintf("\%s\%s", paths$saisonmanager, game_links)

get_scorers(game_links_full[1])

}
