% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_scorers.R
\name{get_teams}
\alias{get_teams}
\title{Get Teams}
\usage{
get_teams(game_link_full)
}
\arguments{
\item{game_link_full}{link to game stats url}
}
\value{
table with team players
}
\description{
Get Teams
}
\examples{
\dontrun{
games_u11 <- get_links_game()[7,]
completed_games_u11 <- urls_completed_games(games_u11$links_url)
if(length(completed_games_u11)>0) get_teams(completed_games_u11[1]) 
}
}
