% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urls_completed_games.R
\name{urls_completed_games}
\alias{urls_completed_games}
\title{URLs of completed games}
\usage{
urls_completed_games(url_games)
}
\arguments{
\item{url_games}{url with games for one group}
}
\value{
vector with urls of completed games
}
\description{
URLs of completed games
}
\examples{
\dontrun{
games_u11 <- get_links_game()[7,]
urls_completed_games(games_u11$links_url)
}
}
