% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_scorers.R
\name{get_scorers}
\alias{get_scorers}
\title{Get Scorers}
\usage{
get_scorers(game_link_full)
}
\arguments{
\item{game_link_full}{link to game stats url}
}
\value{
table with scorers
}
\description{
Get Scorers
}
\examples{
\dontrun{
games_u11 <- get_links_game()[7,]
completed_games_u11 <- urls_completed_games(games_u11$links_url)
if(length(completed_games_u11)>0) get_scorers(completed_games_u11[1])
}
}
