% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_scorers.R
\name{get_teams}
\alias{get_teams}
\title{Get Teams}
\usage{
get_teams(game_html)
}
\arguments{
\item{game_html}{xml2::read_html(game_link_full)  imported html with game stats}
}
\value{
table with team players
}
\description{
Get Teams
}
\examples{
\dontrun{
games_u11 <- get_links_game()[7,]
completed_games_u11 <- urls_completed_games(games_u11$links_url)
if(length(completed_games_u11)>0) get_teams(xml2::read_html(completed_games_u11[1])) 
}
}
